added separate Benchmark console application and restored original Example project
