+++ SevenZipExtractor.sln
@@ -12,6 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Pack.ps1 = Pack.ps1
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Benchmark", "Benchmark\Benchmark.csproj", "{49BCF38D-DAF5-42A5-BC77-B36820FB85FA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,6 +28,10 @@ Global
 		{2EFA3756-853D-4282-A528-C644411112F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2EFA3756-853D-4282-A528-C644411112F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2EFA3756-853D-4282-A528-C644411112F3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{49BCF38D-DAF5-42A5-BC77-B36820FB85FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{49BCF38D-DAF5-42A5-BC77-B36820FB85FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{49BCF38D-DAF5-42A5-BC77-B36820FB85FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{49BCF38D-DAF5-42A5-BC77-B36820FB85FA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
